refactor(racing): add package and change file directories
