+++ src/main/java/car_racing/domain/RacingPlayer.java
@@ -1,4 +1,4 @@
-package car_racing;
+package car_racing.domain;
 
 public interface RacingPlayer {
     void proceedNextTurn(Boolean proceed);
