+++ src/main/java/car_racing/domain/ScoreStrategy.java
@@ -1,4 +1,4 @@
-package car_racing;
+package car_racing.domain;
 
 public interface ScoreStrategy {
     Integer generateScore();
