+++ src/test/java/car_racing/ScoreStrategyTest.java
@@ -1,5 +1,7 @@
 package car_racing;
 
+import car_racing.domain.RandomScoreStrategy;
+import car_racing.domain.ScoreStrategy;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
